refactor variable inputs
