+++ examples/variables.py
@@ -4,46 +4,32 @@
 # http requests to the server
 
 from hera.task import Task
-from hera.input import InputParameterAsEnv
+from hera.variable import VariableAsEnv
 from hera.workflow import Workflow
 from hera.workflow_service import WorkflowService
-
-
-def server():
-    pass
-
-
-def consumer(server_ip):
-    print(server_ip)
-
-
-# TODO: replace the domain and token with your own
-ws = WorkflowService(host='my-argo-server.com', token='my-auth-token')
 
 
 def server():
     from http.server import BaseHTTPRequestHandler, HTTPServer
 
     class MyServer(BaseHTTPRequestHandler):
-        def get_check(self):
+        def do_GET(self):
             self.send_response(200)
             self.send_header("Content-type", "application/json")
             self.end_headers()
             self.wfile.write(bytes("{'name':'John'}", "utf-8"))
 
     webServer = HTTPServer(("0.0.0.0", 8080), MyServer)
-    try:
-        webServer.serve_forever()
-    except KeyboardInterrupt:
-        pass
+    webServer.serve_forever()
 
 
 def consumer():
     import http.client
     import os
 
-    print(os.environ["IP"])
-    connection = http.client.HTTPConnection(f"{os.environ['IP']}:8080")
+    print(os.environ["SERVER_IP"])
+    server_ip = os.environ["SERVER_IP"].replace('"', '')
+    connection = http.client.HTTPConnection(f"{server_ip}:8080")
     connection.request("GET", "/")
     response = connection.getresponse()
     print(response.read())
@@ -54,8 +40,8 @@ namespace = "argo"
 ws = WorkflowService(host='my-argo-server.com', token="token", namespace=namespace)
 w = Workflow('variables', ws)
 d = Task('daemon', server, daemon=True)
-variables = InputParameterAsEnv(name='SERVER_IP', value="{{tasks.daemon.ip}}")
-t = Task('consumer', consumer, variables=[variables])
+server_ip_variable = VariableAsEnv(name='SERVER_IP', value=d.ip)
+t = Task('consumer', consumer, variables=[server_ip_variable])
 
 d >> t
